+++ index.d.ts
@@ -1,0 +1,4 @@
+export * from './src/pageObject';
+export * from '@qavajs/playwright-runner-adapter';
+export * from './src/QavajsPlaywrightWorld';
+export * from './src/QavajsPlaywrightElectronWorld';