+++ src/setup.ts
@@ -1,4 +1,4 @@
-import { setWorldConstructor } from '@cucumber/cucumber';
+import { setWorldConstructor } from '@qavajs/playwright-runner-adapter';
 import { QavajsPlaywrightWorld } from './QavajsPlaywrightWorld';
 
 setWorldConstructor(QavajsPlaywrightWorld);